Start of customer and customer_specs in their own files
