+++ lib/order.rb
@@ -38,9 +38,6 @@ module Grocery
       end
     end
 
-# self means it's related to the class rather than a specific instance
-# we still want to keep this method within the class bc it's related to the class itself (or in other words, it's related to what this code is doing with respect to Order)
-# use self when you want to perform an action within a class, but don't want to do so on one instance at a time...instead want to take some other action that's merely related to the class
     def self.all
       orders_entered = []
       CSV.read("/Users/brandyaustin/ada/week2/grocery_store/grocery-store/support/orders.csv").each do |row|
@@ -68,31 +65,4 @@ module Grocery
     end
 
   end
-
-  class Customer
-
-    attr_reader :id, :email, :address
-
-    def initialize(customer_id, email, address)
-      @id = customer_id
-      @email = email
-      @address = address
-    end
-
-    def self.all
-      customers = []
-      CSV.read("/Users/brandyaustin/ada/week2/grocery_store/grocery-store/support/customers.csv").each do |row|
-        id = row[0].to_i
-        products = {}
-        products_array = row[1].split(";")
-
-        products_array.each do |item|
-          name, price = item.split(':')
-          products[name] = price.to_f
-        end
-        orders_entered << Order.new(id, products)
-      end
-      return orders_entered
-    end
-
 end
